% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDRCData.R
\name{calculateDRCData}
\alias{calculateDRCData}
\title{calculateDRCData}
\usage{
calculateDRCData(idrc_set, cut_time = NULL)
}
\arguments{
\item{idrc_set}{IncucyteDRCSet object}

\item{cut_time}{Desired cut time.  If NULL will use the cut_time in the IncucyteDRCSet object (if set)}
}
\value{
IncucyteDRCSet object
}
\description{
For a given cut time, calculate values from the growth curves in a IncucyteDRCSet object
}
\examples{
test_pm <- importPlatemapXML(system.file(file='extdata/example.PlateMap', package='IncucyteDRC'))
test_data <- importIncucyteData(system.file(file='extdata/example_data.txt', package='IncucyteDRC'), metric='pc')

test_list <- splitIncucyteDRCPlateData(test_pm, test_data, group_columns='growthcondition')

str(test_list)

test_splines <- fitGrowthCurvesIndividual(test_list[[2]])
plotIncucyteDRCSet(test_splines)
test_drc <- calculateDRCData(test_splines, cut_time=100)
names(test_drc)
test_drc$drc_data
plotIncucyteDRCSet(test_drc)

}

