% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEC50.R
\name{calculateEC50}
\alias{calculateEC50}
\title{calculateEC50}
\usage{
calculateEC50(idrc_set, include_control = FALSE)
}
\arguments{
\item{idrc_set}{IncucyteDRCSet object}
}
\value{
IncucyteDRCSet object
}
\description{
Calculates the EC50s for the dose response curves in an IncucyteDRCSet object
}
\examples{
test_pm <- importPlatemapXML(system.file(file='extdata/example.PlateMap', package='IncucyteDRC'))
test_data <- importIncucyteData(system.file(file='extdata/example_data.txt', package='IncucyteDRC'), metric='pc')

test_list <- splitIncucyteDRCPlateData(test_pm, test_data, group_columns='growthcondition')

str(test_list)

test_splines <- fitIndividualSplines(test_list[[2]])
plotIncucyteDRCSet(test_splines)
test_drc <- calculateDRCdata(test_splines, cut_time=100)
plotIncucyteDRCSet(test_drc)
exportDRCtoDataFrame(test_drc)
test_drc <- fitDoseResponseCurve(test_drc)
test_drc <- calculateEC50(test_drc)
exportEC50Data(test_drc)

}

