% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDRCDataToDataFrame.R
\name{exportDRCDataToDataFrame}
\alias{exportDRCDataToDataFrame}
\title{exportDRCDataToDataFrame}
\usage{
exportDRCDataToDataFrame(idrc_set, include_control = FALSE,
  add_metadata = FALSE)
}
\arguments{
\item{idrc_set}{IncucyteDRCSet object}

\item{include_control}{Whether to include control sample as zero conc control}
}
\value{
IncucyteDRCSet object
}
\description{
Exports data into a standard data frame format for an IncucyteDRCSet
}
\examples{
test_pm <- importPlatemapXML(system.file(file='extdata/example.PlateMap', package='IncucyteDRC'))
test_data <- importIncucyteData(system.file(file='extdata/example_data.txt', package='IncucyteDRC'), metric='pc')

test_list <- splitIncucyteDRCPlateData(test_pm, test_data, group_columns='growthcondition')

str(test_list)

test_splines <- fitGrowthCurvesIndividual(test_list[[2]])
plotIncucyteDRCSet(test_splines)
test_drc <- calculateDRCData(test_splines, cut_time=100)
plotIncucyteDRCSet(test_drc)
exportDRCDataToDataFrame(test_drc)

}

