% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDRCDataToPRISM.R
\name{exportDRCDataToPRISM}
\alias{exportDRCDataToPRISM}
\title{exportDRCDataToPRISM}
\usage{
exportDRCDataToPRISM(idrc_set, include_control = FALSE,
  add_metadata = FALSE)
}
\arguments{
\item{idrc_set}{IncucyteDRCSet object}

\item{include_control}{Whether to include control sample as zero conc control}

\item{add_metadata}{Whether or not to merge IncucyteDRCSet metadata into the output}
}
\value{
IncucyteDRCSet object
}
\description{
Exports data in PRISM format for an IncucyteDRCSet
}
\examples{
test_pm <- importPlatemapXML(system.file(file='extdata/example.PlateMap', package='IncucyteDRC'))
test_data <- importIncucyteData(system.file(file='extdata/example_data.txt', package='IncucyteDRC'), metric='pc')

test_list <- splitIncucyteDRCPlateData(test_pm, test_data, group_columns='growthcondition')

str(test_list)

test_splines <- fitGrowthCurvesIndividual(test_list[[2]])
plotIncucyteDRCSet(test_splines)
test_drc <- calculateDRCData(test_splines, cut_time=100)
plotIncucyteDRCSet(test_drc)
exportDRCDataToPRISM(test_drc)

}

