% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportEC50Data.R
\name{exportEC50Data}
\alias{exportEC50Data}
\title{exportEC50Data}
\usage{
exportEC50Data(idrc_set, add_metadata = FALSE)
}
\arguments{
\item{idrc_set}{IncucyteDRCSet object}
}
\value{
IncucyteDRCSet object
}
\description{
Exports EC50 data into a standard data frame format from the dose response fit for an IncucyteDRCSet object
}
\examples{
test_pm <- importPlatemapXML(system.file(file='extdata/example.PlateMap', package='IncucyteDRC'))
test_data <- importIncucyteData(system.file(file='extdata/example_data.txt', package='IncucyteDRC'), metric='pc')

test_list <- splitIncucyteDRCPlateData(test_pm, test_data, group_columns='growthcondition')

str(test_list)

test_drc <- fitIndividualSplines(test_list[[2]])
test_drc <- fitGroupSplines(test_drc)
plotIncucyteDRCSet(test_drc)
test_drc <- calculateCutTime(test_drc)
test_drc <- calculateDRCdata(test_drc)
plotIncucyteDRCSet(test_drc)
exportDRCtoDataFrame(test_drc)
test_drc <- fitDoseResponseCurve(test_drc)
exportEC50Data(test_drc)

}

