% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowthCurvesGrouped.R
\name{fitGrowthCurvesGrouped}
\alias{fitGrowthCurvesGrouped}
\title{fitGrowthCurvesGrouped}
\usage{
fitGrowthCurvesGrouped(idrc_set)
}
\arguments{
\item{idrc_set}{IncucyteDRCSet object}
}
\value{
IncucyteDRCSet object
}
\description{
Function to fit loess function to the growth curve data in an IncucyteDRCSet object
}
\examples{
test_pm <- importPlatemapXML(system.file(file='extdata/example.PlateMap', package='IncucyteDRC'))
test_data <- importIncucyteData(system.file(file='extdata/example_data.txt', package='IncucyteDRC'), metric='pc')

test_list <- splitIncucyteDRCPlateData(test_pm, test_data, group_columns='growthcondition')

str(test_list)

test_splines <- fitGrowthCurvesGrouped(test_list[[2]])

}

