% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowthCurvesIndividual.R
\name{fitGrowthCurvesIndividual}
\alias{fitGrowthCurvesIndividual}
\title{fitGrowthCurvesIndividual}
\usage{
fitGrowthCurvesIndividual(idrc_set, span = 0.3)
}
\arguments{
\item{idrc_set}{IncucyteDRCSet object}

\item{span}{Set the span to be used in the curve fit.  Default 0.3.}
}
\value{
IncucyteDRCSet object
}
\description{
Function to fit splines to the growth curve data in an IncucyteDRCSet object
}
\examples{
pm_file <- system.file(file='extdata/example.PlateMap', package='IncucyteDRC')
test_pm <- importPlatemapXML(pm_file)
data_file <- system.file(file='extdata/example_data.txt', package='IncucyteDRC')
test_data <- importIncucyteData(data_file, metric='pc')

test_list <- splitIncucyteDRCPlateData(test_pm, test_data, group_columns='growthcondition')

print(test_list)

test_idrc_set <- fitGrowthCurvesIndividual(test_list[[2]])

}

