% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importIncucyteData.R
\name{importIncucyteData}
\alias{importIncucyteData}
\title{importIncucyteData}
\usage{
importIncucyteData(filepath, metric = "pc", plateid = basename(filepath))
}
\arguments{
\item{filepath}{Path to a data file generated by the Incucyte Zoom software}

\item{metric}{Metric must be either pc (percent confluence) or ca (confluence area).  Default is pc.}

\item{plateid}{An identifier for the plate.  If set to NULL then filename used as default}
}
\value{
IncucyteDRCPlateData object
}
\description{
Imports data exported from Incucyte Zoom software-  NEED MORE DOCUMENTATION HERE ABOUT HOW TO DO EXPORT!!
}
\examples{
test_data <- importIncucyteData(system.file(file='extdata/example_data.txt', package='IncucyteDRC'), metric='pc')
str(test_data)
head(test_data$data)
}

