% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncucyteDRC_documentation.R
\docType{package}
\name{IncucyteDRC-package}
\alias{IncucyteDRC}
\alias{IncucyteDRC-package}
\title{Dose Response Curves from Incucyte Proliferation Assays}
\description{
To open the vignette type \code{vignette("Overview")}
More info to be added...
}
\author{
Phil Chapman <phil.chapman@cruk.manchester.ac.uk>
}
\references{
See F1000R article
}

