% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlatemap.R
\name{plotPlatemap}
\alias{plotPlatemap}
\title{plotPlatemap}
\usage{
plotPlatemap(pm)
}
\arguments{
\item{df}{Data frame containing platemap data}
}
\value{
ggplot object
}
\description{
Visualisation of an Incucyte platemap
}
\examples{
test_pm <- importPlatemapXML(system.file(file='extdata/example.PlateMap', package='IncucyteDRC'))
plotPlatemap(test_pm)

library(magrittr)
importPlatemapXML(system.file(file='extdata/example2.PlateMap', package='IncucyteDRC')) \%>\%
  plotPlatemap()
}

