% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateIncucyteDRCSetMetadata.R
\name{populateIncucyteDRCSetMetadata}
\alias{populateIncucyteDRCSetMetadata}
\title{populateIncucyteDRCSetMetadata}
\usage{
populateIncucyteDRCSetMetadata(idrc_set, group_columns)
}
\arguments{
\item{idrc_set}{IncucyteDRCSet object}

\item{group_columns}{Vector of columns that are present in the data frame to generate the groups.}
}
\value{
IncucyteDRCSet object
}
\description{
Function to populate the metadata slot of an IncucyteDRCSet object.  Not exported, used at part of the
splitIncucyteDRCPlateData function.
}

