% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printIncucyteDRCPlateData.R
\name{print.IncucyteDRCPlateData}
\alias{print.IncucyteDRCPlateData}
\title{print.IncucyteDRCPlateData}
\usage{
\method{print}{IncucyteDRCPlateData}(x)
}
\arguments{
\item{x}{IncucyteDRCPlateData object}
}
\description{
Prints information on an IncucyteDRCPlateData object
}
\examples{
test_data <- importIncucyteData(system.file(file='extdata/example_data.txt', package='IncucyteDRC'), metric='pc')
print(test_data)
}

