% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printIncucyteDRCSet.R
\name{print.IncucyteDRCSet}
\alias{print.IncucyteDRCSet}
\title{print.IncucyteDRCSet}
\usage{
\method{print}{IncucyteDRCSet}(x)
}
\arguments{
\item{x}{IncucyteDRCSet object}
}
\description{
Prints information on an IncucyteDRCSet object
}
\examples{
test_pm <- importPlatemapXML(system.file(file='extdata/example.PlateMap', package='IncucyteDRC'))
test_data <- importIncucyteData(system.file(file='extdata/example_data.txt', package='IncucyteDRC'), metric='pc')
test_list <- splitIncucyteDRCPlateData(test_pm, test_data, group_columns='growthcondition')
class(test_list)
class(test_list[[2]])
print(test_list[[2]])
}

