% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVisServer.R
\name{shinyVisServer}
\alias{shinyVisServer}
\title{Shiny Visualisation Server}
\usage{
shinyVisServer(input, output)
}
\arguments{
\item{input}{Shiny input list}

\item{output}{Shiny output list}
}
\value{
A shiny server
}
\description{
Creates a shiny server for the interactive web app for the running the IncucyteDRC workflow
}

