% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVisUI.R
\name{shinyVisUI}
\alias{shinyVisUI}
\title{Shiny Visualisation UI}
\usage{
shinyVisUI()
}
\value{
A shiny UI
}
\description{
Creates a shiny UI for the interactive web app for the running the IncucyteDRC workflow
}

