% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importIncucyteData.R
\name{importIncucyteData}
\alias{importIncucyteData}
\title{importIncucyteData}
\usage{
importIncucyteData(filepath, metric = "pc", plateid = basename(filepath))
}
\arguments{
\item{filepath}{Path to a data file generated by the Incucyte Zoom software}

\item{metric}{Metric must be either pc (percent confluence) or ca (confluence area).  Default is pc.}

\item{plateid}{An identifier for the plate.  If set to NULL then filename used as default}
}
\value{
IncucyteDRCPlateData object
}
\description{
Imports data exported from Incucyte Zoom software-  NEED MORE DOCUMENTATION HERE ABOUT HOW TO DO EXPORT!!
}
\examples{
#dataset 1
data_file1 <- system.file(file='extdata/example_data.txt', package='IncucyteDRC')
test_data1 <- importIncucyteData(data_file1, metric='pc')
test_data1
head(test_data1$data)

#dataset 2
data_file2 <- system.file(file='extdata/example_data2.txt', package='IncucyteDRC')
test_data2 <- importIncucyteData(data_file2, metric='pc')
}

