% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPlatemapXML.R
\name{importPlatemapXML}
\alias{importPlatemapXML}
\title{importPlatemapXML}
\usage{
importPlatemapXML(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
data fame
}
\description{
Imports .Platemap XML files from Incucyte Zoom software
}
\examples{
test_df <- importPlatemapXML(system.file(file='extdata/example.PlateMap', package='IncucyteDRC'))
head(test_df)
}

