% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDoseResponseCurve.R
\name{plotDoseResponseCurve}
\alias{plotDoseResponseCurve}
\title{plotDoseResponseCurve}
\usage{
plotDoseResponseCurve(idrc_set, sampleid)
}
\arguments{
\item{idrc_set}{IncucyteDRCSet object}

\item{sampleid}{The sample id to plot}
}
\value{
a plot
}
\description{
Plots the dose response curve for a given sample id from an IncucyteDRCSet object
}
\examples{
test_pm <- importPlatemapXML(system.file(file='extdata/example.PlateMap', package='IncucyteDRC'))
test_data <- importIncucyteData(system.file(file='extdata/example_data.txt', package='IncucyteDRC'), metric='pc')

test_list <- splitIncucyteDRCPlateData(test_pm, test_data, group_columns='growthcondition')

str(test_list)

test_drc <- fitGrowthCurvesIndividual(test_list[[2]])
test_drc <- fitGrowthCurvesGrouped(test_drc)
plotIncucyteDRCSet(test_drc)
test_drc <- calculateCutTimeForIDRCSet(test_drc)
test_drc <- calculateDRCData(test_drc)
test_drc <- fitDoseResponseCurve(test_drc)
test_drc <- calculateEC50(test_drc)
plotDoseResponseCurve(test_drc, 'PDD00017273')
}

