% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIncucyteDRCSet.R
\name{plotIncucyteDRCSet}
\alias{plotIncucyteDRCSet}
\title{plotIncucyteDRCSet}
\usage{
plotIncucyteDRCSet(idrc_set, grouped = FALSE)
}
\arguments{
\item{idrc_set}{IncucyteDRCSet object}
}
\value{
ggplot object
}
\description{
Plot the growth curve data from an IncucyteDRCSet object
}
\examples{
test_pm <- importPlatemapXML(system.file(file='extdata/example.PlateMap', package='IncucyteDRC'))
test_data <- importIncucyteData(system.file(file='extdata/example_data.txt', package='IncucyteDRC'), metric='pc')

test_list <- splitIncucyteDRCPlateData(test_pm, test_data, group_columns='growthcondition')

str(test_list)

test_splines <- fitGrowthCurvesIndividual(test_list[[2]])
test_splines <- fitGrowthCurvesGrouped(test_splines)
plotIncucyteDRCSet(test_splines, grouped=FALSE)
plotIncucyteDRCSet(test_splines, grouped=TRUE)

}

