% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlatemap.R
\name{plotPlatemap}
\alias{plotPlatemap}
\title{plotPlatemap}
\usage{
plotPlatemap(pm)
}
\arguments{
\item{pm}{Platemap data frame geneated by either importPlatemap or importPlatemapXML functions.}
}
\value{
ggplot object
}
\description{
Visualisation of an Incucyte platemap
}
\examples{
pm_file <- system.file(file='extdata/example.PlateMap', package='IncucyteDRC')
test_pm <- importPlatemapXML(pm_file)
plotPlatemap(test_pm)

library(magrittr)
pm_file2 <- system.file(file='extdata/example2.PlateMap', package='IncucyteDRC')
importPlatemapXML(pm_file2) \%>\% plotPlatemap()
}

