% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printIncucyteDRCPlateData.R
\name{print.IncucyteDRCPlateData}
\alias{print.IncucyteDRCPlateData}
\title{print.IncucyteDRCPlateData}
\usage{
\method{print}{IncucyteDRCPlateData}(x, ...)
}
\arguments{
\item{x}{IncucyteDRCPlateData object}

\item{...}{Additional arguments}
}
\description{
Prints information on an IncucyteDRCPlateData object
}
\examples{
data_file <- system.file(file='extdata/example_data.txt', package='IncucyteDRC')
test_data <- importIncucyteData(data_file, metric='pc')
print(test_data)
}

