% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printIncucyteDRCSetList.R
\name{print.IncucyteDRCSetList}
\alias{print.IncucyteDRCSetList}
\title{print.IncucyteDRCSetList}
\usage{
\method{print}{IncucyteDRCSetList}(x, ...)
}
\arguments{
\item{x}{IncucyteDRCSet object}

\item{...}{Additional arguments}
}
\description{
Prints information on an IncucyteDRCSetList object
}
\examples{
pm_file <- system.file(file='extdata/example.PlateMap', package='IncucyteDRC')
test_pm <- importPlatemapXML(pm_file)
data_file <- system.file(file='extdata/example_data.txt', package='IncucyteDRC')
test_data <- importIncucyteData(data_file, metric='pc')
test_list <- splitIncucyteDRCPlateData(test_pm, test_data, group_columns='growthcondition')
class(test_list)
print(test_list)
}

