% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVisApp.R
\name{shinyVisApp}
\alias{shinyVisApp}
\title{Shiny Visualisation App}
\usage{
shinyVisApp()
}
\value{
Launches an interactive Shiny application
}
\description{
Provides an interactive web app for the running the IncucyteDRC workflow
}

